pyproject.toml
[project]
name = "bentoml-local-setup"
version = "0.1.0"
description = "BentoML local setup for macOS without Docker"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "bentoml[io]>=1.2.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "prometheus-client>=0.15.0",
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0",
    "diffusers>=0.21.0",
    "transformers>=4.25.0",
    "accelerate>=0.20.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=9.0.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.0.0",
    "llama-cpp-python>=0.2.0",
    "mlx-whisper>=0.1.0",
    "pydub>=0.25.0",
    "requests>=2.28.0",
    "orjson>=3.8.0",
    "realesrgan>=0.3.0",
    "gfpgan>=1.3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
    "notebook>=6.5.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
    "notebook>=6.5.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["bentoml"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings  
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "timeout: marks tests with custom timeout values",
]
addopts = [
    "--strict-markers",
    "--disable-warnings", 
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--timeout=30",
    "--timeout-method=thread",
]
pythonpath = ["."]
timeout = 30
timeout_method = "thread"
